chore: fix pre-existing nightly clippy lints
These two lints are unrelated to the fullscreen effect changes but
block CI for this PR. Both lines exist on upstream/main and were
introduced years ago; newer nightly clippy versions started flagging
them. Pinning the toolchain is out of scope for this PR, so fixing
them in place is the lowest-friction path to a green check.

- packages/wm/src/user_config.rs:86 - useless_borrows_in_formatting
- packages/wm-platform/src/platform_impl/windows/mouse_listener.rs:189 - unnecessary_cast

+++ packages/wm-platform/src/platform_impl/windows/mouse_listener.rs
@@ -182,11 +182,10 @@ impl MouseListener {
     // Ignore if input is invalid or not a mouse event. Inputs from our own
     // process are also ignored, since `NativeWindow::focus` simulates
     // mouse input.
-    #[allow(clippy::cast_possible_truncation)]
     if res_size == 0
       || raw_input_size == u32::MAX
       || raw_input.header.dwType != RIM_TYPEMOUSE.0
-      || unsafe { raw_input.data.mouse.ulExtraInformation } as u32
+      || unsafe { raw_input.data.mouse.ulExtraInformation }
         == FOREGROUND_INPUT_IDENTIFIER
     {
       return Ok(());
